fix rblquery_t queryid truncation of uint32_t DNS ID to uint16_t

+++ src/s_auth.c
@@ -90,7 +90,7 @@ typedef struct _rblquery
         rb_dlink_node node;
         struct AuthRequest *auth;
         rbl_t *rbl;
-        uint16_t queryid;
+        uint32_t queryid;
 } rblquery_t;
